backend/requirements.txt
fastapi
uvicorn
SQLModel
pydantic-settings
passlib[bcrypt]
resend